fix input form errors

+++ src/controllers/activities.controller.ts
@@ -4,6 +4,7 @@ import { ActivityService } from "../services/activities.service";
 import { ActivityOutput, ActivityCreateInput, ActivityUpdateInput, ActivityResponse, ActivityCreateRequest, ActivityQueryRequest, ActivityUpdateRequest, ActivityQueryInput } from '../types/activities.type';
 import { ApiResponse } from "../types/controller.type";
 import { ActivityLogsService } from "../services/activitylogs.service";
+import AppError from "../types/error.type";
 
 export class ActivityController {
     private service = new ActivityService();
@@ -15,7 +16,7 @@ export class ActivityController {
             subcategoryid: req.body.subcategoryid,
             trackingtypeid: req.body.trackingtypeid,
             userid: req.user!.id,
-            minvalue:req.body.minvalue
+            minvalue: req.body.minvalue
         }
         const result = await this.service.create(input);
         const output: ActivityResponse = {
@@ -23,7 +24,7 @@ export class ActivityController {
             name: result.name,
             subcategoryid: result.subcategoryid,
             trackingtypeid: result.trackingtypeid,
-            minvalue:result.minvalue
+            minvalue: result.minvalue
         };
         res.status(200).json({
             success: true,
@@ -35,16 +36,16 @@ export class ActivityController {
 
     findById = async (req: Request, res: Response<ApiResponse<ActivityResponse>>, next: NextFunction): Promise<void> => {
         const result = await this.service.findById(Number(req.params.id), req.user!.id);
-        const subResult = await this.subService.findAll({activityid:Number(req.params.id),userid:req.user!.id});
+        const subResult = await this.subService.findAll({ activityid: Number(req.params.id), userid: req.user!.id });
         if (result) {
             const output: ActivityResponse = {
                 id: result.id,
                 name: result.name,
                 subcategoryid: result.subcategoryid,
                 trackingtypeid: result.trackingtypeid,
-                minvalue:result.minvalue,
-                trackingtypename:result.trackingtypename,
-                data:subResult
+                minvalue: result.minvalue,
+                trackingtypename: result.trackingtypename,
+                data: subResult
             }
             res.status(200).json({
                 success: true,
@@ -83,18 +84,18 @@ export class ActivityController {
         const input: ActivityUpdateInput = {
             name: req.body.name,
             userid: req.user!.id,
-            minvalue:req.body.minvalue,
-            subcategoryid:req.body.subcategoryid!,
-            trackingtypeid:req.body.trackingtypeid!
+            minvalue: req.body.minvalue,
+            subcategoryid: req.body.subcategoryid!,
+            trackingtypeid: req.body.trackingtypeid!
         }
         const result = await this.service.update(Number(req.params.id), input);
 
-        const output:ActivityResponse = {
-            id:result.id,
-            name:result.name,
-            subcategoryid:result.subcategoryid,
-            trackingtypeid:result.trackingtypeid,
-            minvalue:result.minvalue
+        const output: ActivityResponse = {
+            id: result.id,
+            name: result.name,
+            subcategoryid: result.subcategoryid,
+            trackingtypeid: result.trackingtypeid,
+            minvalue: result.minvalue
         }
 
         res.status(200).json({
@@ -106,6 +107,10 @@ export class ActivityController {
     }
 
     delete = async (req: Request, res: Response<ApiResponse<boolean>>, next: NextFunction): Promise<void> => {
+        const subResult = await this.subService.findAll({ activityid: Number(req.params.id), userid: req.user!.id });
+        if (subResult.length > 0) {
+            throw new AppError(400, "Activity has items!");
+        }
         const result = await this.service.delete(Number(req.params.id), req.user!.id);
         res.status(200).json({
             success: true,
