+++ src/controllers/subcategories.controller.ts
@@ -4,6 +4,7 @@ import { SubCategoryService } from "../services/subcategories.service";
 import { SubCategoryOutput, SubCategoryCreateInput, SubCategoryUpdateInput, SubCategoryResponse, SubCategoryCreateRequest, SubCategoryQueryRequest, SubCategoryUpdateRequest, SubCategoryQueryInput } from '../types/subcategories.type';
 import { ApiResponse } from "../types/controller.type";
 import { ActivityService } from "../services/activities.service";
+import AppError from "../types/error.type";
 
 export class SubCategoryController {
     private service = new SubCategoryService();
@@ -92,6 +93,10 @@ export class SubCategoryController {
     }
 
     delete = async (req: Request, res: Response<ApiResponse<boolean>>, next: NextFunction): Promise<void> => {
+        const subResult = await this.subService.findAll({ subcategoryid: Number(req.params.id), userid: req.user!.id });
+        if(subResult.length > 0){
+            throw new AppError(400,"Subcategory has items!");
+        }
         const result = await this.service.delete(Number(req.params.id), req.user!.id);
         res.status(200).json({
             success: true,
